Fully modernize the HelloAndroidCamera2 build system
Same build-system rot as HelloAndroid: an unbuildable Gradle 2.2/AGP
1.2.2 setup resolving from the dead jcenter(), and a src/res layout
predating Android Studio. Replace it with the standard
app/src/main/{java,res,cpp} module layout and a single
app/src/main/cpp/CMakeLists.txt driven by Gradle's externalNativeBuild,
adapted for this app's two generators (deinterleave, edge_detect) and
four JNI wrapper sources. Targets arm64-v8a and x86_64.

Also fixes three runtime issues: CameraActivity never requested the
CAMERA permission at runtime (added the standard
ContextCompat/ActivityCompat check-then-request flow);
Camera2BasicFragment.java's `switch (view.getId()) { case R.id.toggle:
}` doesn't compile under AGP's non-final resource IDs default, rewritten
as an if; and YuvBufferT's "packed planar" chroma case built a Buffer
for just the U (or V) plane and then called crop(1, 0, height * 2) to
try to extend it to cover both planes -- Buffer::crop() only ever
narrows within a buffer's existing declared bounds, so this crashed
(SIGABRT) the moment the camera produced a frame in this layout. Fixed
by constructing packedPlanarChromaView_ directly from the pointer with
the full doubled-height shape, the same way the sibling "interleaved"
branch already does.

Fixes #8728
