+++ apps/HelloAndroidCamera2/.gitignore
@@ -1,8 +1,8 @@
-.gradle/**
-gen/**
-gradle_build/**
-*.iml
+.gradle/
+build/
+app/build/
+app/.cxx/
+app/.cxx-generators/
 local.properties
-obj/**
-proguard-project.txt
-project.properties
+.idea/
+*.iml
