+++ apps/HelloAndroidCamera2/README.md
@@ -20,75 +20,35 @@ a logcat error message that looks something like:
 E/halide_native( 6146): ANativeWindow buffer locked but its size was 1920 x
 1440, expected 1440 x 1080
 
-This application builds for multiple native ABIs. (At present armeabi,
-armeabi-v7a, arm64-v8a, x86_64, and x86 are supported.) Halide code is generated
-for each architecture.
+This app targets `arm64-v8a` and `x86_64`. The two Halide pipelines
+(`app/src/main/cpp/generators/deinterleave_generator.cpp`,
+`edge_detect_generator.cpp`) are compiled once per ABI via Gradle's
+`externalNativeBuild`, which drives a single CMake project
+(`app/src/main/cpp/CMakeLists.txt`); the Halide generators themselves are built
+with the host compiler as a side effect of that same CMake configure, so no
+separate pre-build step is needed.
 
-This build is meant to use Android command line tools. (An IDE is not required.)
-In order to build, the following will be required:
+## Prerequisites
 
-- Android NDK -- This can be downloaded here:
-  https://developer.android.com/tools/sdk/ndk/index.html After installing, make
-  sure the top-level directory of the install is in the PATH. (It should contain
-  an executable ndk-build file.)
+- A local Halide install (built via CMake -- see the top-level
+  `doc/BuildingHalideWithCMake.md`).
+- Android Studio, or the Android SDK + NDK (Android Studio's SDK Manager can
+  install the NDK under SDK Tools -- "NDK (Side by side)").
+- A CMake >= 3.28. The CMake bundled with the Android SDK by default is often
+  older than this; if so, point Gradle at a newer one by adding
+  `cmake.dir=<path to a CMake >= 3.28 install prefix>` to `local.properties`
+  (e.g. `cmake.dir=/opt/homebrew` for a Homebrew-installed CMake on macOS).
 
-- Android SDK -- This can be downloaded here:
-  http://developer.android.com/sdk/index.html The standalone SDK is desired.
-  Once downloaded, the "android" program in the tools directory of the install
-  will need to be run. It should bring up a UI allowing one to choose components
-  to install. HelloAndroidCamera2 currently depends on the android-21 release.
-  Make sure the tools directory is on one's PATH.
-
-- Apache Ant -- which can be downloaded here:
-  http://ant.apache.org/bindownload.cgi make sure the bin directory is on one's
-  PATH.
-
-If everything is setup correctly, running the build.sh script in this directory,
-with the current directory set to here, would build the HelloAndroidCamera2 apk
-and install it on a connected Android device.
-
-# Gradle
-
-To use Gradle create local.properties file in this folder with sdk.dir and
-ndk.dir variables defined like so:
+## Building
 
 ```
-sdk.dir=/Users/joe/Downloads/android-sdk
-ndk.dir=/Users/joe/Downloads/android-ndk
+./gradlew assembleDebug -PHalide_ROOT=<path/to/halide install>
 ```
 
-After that run `gradlew build` which will produce .apk file ready for deployment
-to the Android device.
+(`Halide_ROOT` can also be provided via an environment variable of the same
+name.) This produces `app/build/outputs/apk/debug/app-debug.apk`, installable
+with `adb install`.
 
-On Linux/Mac you can use `build-gradle.sh` to build, deploy and run this sample
-application.
-
-Pay attention to the list of platforms supported by your Halide installation.
-They are listed in jni/Application.mk APP_ABI variable and in build.gradle archs
-map. For example, if your Halide installation was built without arm64-v8a,
-remove it from APP_ABI and archs. Both list and map should match, otherwise you
-will be getting compilation errors complaining about a missing
-halide_generated.h file:
-
-```
-:ndkBuild FAILED
-
-FAILURE: Build failed with an exception.
-
-* What went wrong:
-Execution failed for task ':ndkBuild'.
-...
-  Output:
-    /private/tmp/7/halide/apps/HelloAndroidCamera2/jni/native.cpp:11:26: fatal error: deinterleave.h: No such file or directory
-     #include "deinterleave.h"
-
-```
-
-# Android Studio
-
-To load project into Android Studio use "File/Import Project..." in Android
-Studio and point to apps/HelloAndroidCamera2/build.gradle file.
-
-You will have to edit automatically-generated local.properties file to add
-ndk.dir property so it points to your Android NDK installation as described in
-Gradle section above.
+To build in Android Studio instead, open this directory as a project, add
+`Halide_ROOT=<path/to/halide install>` to `gradle.properties` (or
+`local.properties`), and use the Run button.
