+++ apps/HelloAndroidCamera2/app/src/main/cpp/AndroidBufferUtilities.h
