+++ apps/HelloAndroidCamera2/app/src/main/cpp/HalideFilters.cpp
