+++ apps/HelloAndroidCamera2/app/src/main/cpp/LockedSurface.cpp
