+++ apps/HelloAndroidCamera2/app/src/main/cpp/generators/edge_detect_generator.cpp
