+++ apps/HelloAndroidCamera2/app/src/main/java/com/android/ex/camera2/utils/StateWaiter.java
