+++ apps/HelloAndroidCamera2/app/src/main/java/com/example/helloandroidcamera2/AutoFitSurfaceView.java
