+++ apps/HelloAndroidCamera2/app/src/main/res/drawable-hdpi/ic_launcher.png
