+++ apps/HelloAndroidCamera2/app/src/main/res/drawable-ldpi/ic_launcher.png
