+++ apps/HelloAndroidCamera2/app/src/main/res/drawable-mdpi/ic_launcher.png
