+++ apps/HelloAndroidCamera2/app/src/main/res/drawable-xhdpi/ic_launcher.png
