+++ apps/HelloAndroidCamera2/app/src/main/res/layout/activity_camera.xml
