+++ apps/HelloAndroidCamera2/app/src/main/res/layout/fragment_camera2_basic.xml
