+++ apps/HelloAndroidCamera2/app/src/main/res/layout/main.xml
