+++ apps/HelloAndroidCamera2/app/src/main/res/values/strings.xml
