+++ apps/HelloAndroidCamera2/app/src/main/res/values/styles.xml
