+++ apps/HelloAndroidCamera2/gradle.properties
@@ -1,0 +1,2 @@
+org.gradle.jvmargs=-Xmx2048m
+android.useAndroidX=true
