--- apps/HelloAndroidCamera2/jni/Application.mk
@@ -1,4 +1,0 @@
-APP_ABI := armeabi armeabi-v7a arm64-v8a x86_64 x86
-APP_PLATFORM := android-21
-APP_STL := c++_static
-APP_CPPFLAGS := -std=c++17 -fno-rtti -fexceptions
